# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Träwelling API",
        "description": "Träwelling user API description. This is an incomplete documentation with still many errors. The API is currently not yet stable. Endpoints are still being restructured. Both the URL and the request or body can be changed. Breaking changes will be announced on GitHub: https://github.com/Traewelling/traewelling/blob/develop/API_CHANGELOG.md",
        "contact": {
            "email": "support@traewelling.de"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0 - alpha"
    },
    "servers": [
        {
            "url": "https://traewelling.de/api/v1",
            "description": "Production Server"
        },
        {
            "url": "http://localhost:8000/api/v1",
            "description": "This instance"
        }
    ],
    "paths": {
...
        "/user/{user}/trusted": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get all trusted users for a user",
                "description": "Get all trusted users for the current user or a specific user (admin only).",
                "operationId": "trustedUserIndex",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID of the user (or string 'self' for current user)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of trusted users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TrustedUserResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Add a user to the trusted users for a user",
                "description": "Add a user to the trusted users for the current user or a specific user (admin only).",
                "operationId": "trustedUserStore",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID of the user (or string 'self' for current user) who want's to trust.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "userId": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "expiresAt": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-07-28T00:00:00Z",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User added to trusted users"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/user/self/trusted-by": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get all users who trust the current user",
                "operationId": "trustedByUserIndex",
                "responses": {
                    "200": {
                        "description": "List of users who trust the current user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TrustedUserResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/user/{user}/trusted/{trusted}": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Remove a user from the trusted users for a user",
                "description": "Remove a user from the trusted users for the current user or a specific user (admin only).",
                "operationId": "trustedUserDestroy",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID of the user (or string 'self' for current user)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "trusted",
                        "in": "path",
                        "description": "ID of the trusted user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "User removed from trusted users"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "LightUserResource": {
                "title": "LightUser",
                "description": "User model with just basic information",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "displayName": {
                        "type": "string",
                        "example": "Gertrud"
                    },
                    "username": {
                        "type": "string",
                        "example": "Gertrud123"
                    },
                    "profilePicture": {
                        "type": "string",
                        "example": "https://traewelling.de/@Gertrud123/picture"
                    },
                    "mastodonUrl": {
                        "type": "string",
                        "example": "https://traewelling.social/@Gertrud123"
                    },
                    "preventIndex": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
...
            "TrustedUserResource": {
                "title": "TrustedUser",
                "properties": {
                    "user": {
                        "$ref": "#/components/schemas/LightUserResource"
                    },
                    "expiresAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-28T00:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "OAuth2 authorizationCode Flow. Get your token from https://traewelling.de/settings/applications. Set the redirect URL to https://traewelling.de/api/oauth2-callback to use it here. See also https://laravel.com/docs/9.x/passport",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "authorizationCode": {
                        "authorizationUrl": "http://localhost:8000/oauth/authorize",
                        "tokenUrl": "http://localhost:8000/oauth/token",
                        "refreshUrl": "http://localhost:8000/auth/refresh",
                        "scopes": {
                            "read-statuses": "see all statuses",
                            "read-notifications": "see your notifications",
                            "read-statistics": "see your statistics",
                            "read-search": "search in Träwelling",
                            "write-statuses": "create, edit, delete statuses",
                            "write-likes": "create and remove likes",
                            "write-notifications": "mark notifications as read, clear notifications",
                            "write-exports": "request data exports",
                            "write-follows": "follow and unfollow users",
                            "write-followers": "accept follow requests and remove followers",
                            "write-blocks": "block and unblock users, mute and unmute users",
                            "write-event-suggestions": "suggest events in your name",
                            "write-support-tickets": "create support tickets in your name",
                            "read-settings": "see your settings, email, etc.",
                            "write-settings-profile": "edit your profile",
                            "read-settings-profile": "see your profile data, e.g. email",
                            "write-settings-mail": "edit your email",
                            "write-settings-profile-picture": "edit your profile picture",
                            "write-settings-privacy": "change your privacy settings",
                            "read-settings-followers": "show follow-requests and followers",
                            "write-settings-calendar": "create and delete new calendar-tokens",
                            "extra-write-password": "change your password",
                            "extra-terminate-sessions": "log you out of other sessions and apps",
                            "extra-delete": "delete your account"
                        }
                    }
                }
            },
            "token": {
...
    },
...
}